fix #41765
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@238237 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/recode/recode.c
@@ -132,7 +132,7 @@ PHP_FUNCTION(recode_string)
 {
 	RECODE_REQUEST request = NULL;
 	char *r = NULL;
-	int r_len = 0, r_alen = 0;
+	size_t r_len = 0, r_alen = 0;
 	int req_len, str_len;
 	char *req, *str;
